Problem: No way to relay fault information to frontend
The source of this problem is that the version of Uptight we were depending on didn't have a way to encode *any* exception in ToJSON instance of Uptight.Trace.
It caused crashes of Jason.encode(). It meant that we need a generic encoding for any exception.
In the latest commit of Uptight, we have elected to encode Elixir exception terms in a best-effort fashion.

Solution:

 - Uptdate the dependency to the new Uptight version.