Problem: No way to generate commanded DBs
Solution:

To add it, follow @maslowalex's instructions and add the needed mix recipe calls to Makefile.
For it to work, however, we needed to significantly improve and generalise create-role-phoenix.sh script and remove file manipulations from Makefile.
Indeed, now create-role-phoenix.sh and only it manipulates config files, it also respects ~/.pgpass and doesn't mess with it at all.
It has an unpleasant side-effect, which is if you use this pgsql bootstrapping script in one project, and then forget to turn off the DB in another project and provision DB from there without changing the port from default 5666 to something else.
Eventually we want to randomise the port, but for the time being it's ok as is.